migrate-dp-attn-mixin-move: Move DP-attention adapter methods to SchedulerDPAttnAdapter
Mechanical cut + paste for the ``migrate-dp-attn-mixin`` mech move.

Cut ``prepare_mlp_sync_batch`` / ``maybe_prepare_mlp_sync_batch`` /
``get_idle_batch`` (@staticmethods after prep) from
``scheduler_dp_attn_mixin.py`` and paste them into ``SchedulerDPAttnAdapter``
class body in ``scheduler_components/dp_attn.py``.

Module-level ``MLPSyncBatchInfo`` dataclass + ``_update_gather_batch`` +
``prepare_mlp_sync_batch_raw`` are cut verbatim from the old mixin and
prepended into the new target module above the adapter class. The source
file is deleted, the ``SchedulerDPAttnMixin`` entry is dropped from the
Scheduler inheritance list, and its ``from`` import is dropped from
``scheduler.py``.

Method bodies otherwise byte-identical. ``@staticmethod`` decorators
dropped; ``self: "SchedulerDPAttnAdapter"`` annotation simplified to bare
``self``.

All callers updated:
  ``self.maybe_prepare_mlp_sync_batch(self.dp_attn_adapter, batch, ...)`` →
  ``self.dp_attn_adapter.maybe_prepare_mlp_sync_batch(batch, ...)``
(pure prefix transformation).

Refactor chain ID: migrate-dp-attn-mixin-move
